Frames

Frames

Overview

Functions

Overview

The most important frame conversion when dealing with the Circular Restricted Three-Body Problem (CR3BP) is between the rotating and inertial frames.

Functions

ThreeBodyProblem.rot2inert — Function

rot2inert(rv, θ, μ; origin=:barycenter)

Convert state vector rv = [r; v] {NON; NON} from rotating (synodic) frame to inertial frame in the normalized CR3BP given time θ {NON} and mass parameter μ {NON}.

rot2inert(rv, θ, sys::System; origin=:barycenter)

Convert state vector rv = [r; v] {NON; NON} from rotating (synodic) frame to inertial frame in the normalized CR3BP given time θ {NON} and CR3BP system sys.

rot2inert(rv, θ, p::Array; origin=:barycenter)

Convert state vector rv = [r; v] {NON; NON} from rotating (synodic) frame to inertial frame in the non-normalized CR3BP given time θ {NON} and p = [μ₁;μ₂;d]` {km³/s²; km³/s²; km}, which contains the gravitational parameters of the first and second primary bodies as well as the distance between them.

ThreeBodyProblem.rot2inert! — Function

rot2inert!(rv, θ, μ)

In-place version of rot2inert(rv, θ, μ)

rot2inert!(rv, θ, sys::System)

In-place version of rot2inert(rv, θ, sys::System)

rot2inert!(rv, θ, p::Array)

In-place version of rot2inert(rv, θ, p::Array)

ThreeBodyProblem.inert2rot — Function

inert2rot(rv, θ, μ)

Inertial frame to rotating (synodic) frame

inert2rot(rv, θ, sys::System)

Inertial frame to rotating (synodic) frame

inert2rot(rv, θ, p::Array)

Inertial frame to rotating (synodic) frame

ThreeBodyProblem.inert2rot! — Function

inert2rot!(rv, t, μ)

In-place version of inert2rot(rv, t, μ).

inert2rot!(rv, t, sys::System)

In-place version of inert2rot(rv, t, sys::System).

inert2rot!(rv, t, p::Array)

In-place version of inert2rot(rv, t, p::Array).

ThreeBodyProblem.enu2ecef — Function

enu2ecef(rv_enu, ϕ, λ, h=0; geodetic=true, ang_unit=:deg)

Converts state from ENU frame to inertial frame

ThreeBodyProblem.ecef2enu — Function

ecef2enu(rv_ecef, ϕ, λ, h=0; geodetic=true, ang_unit::Symbol=:deg, e_earth=0.0818, r_earth=6.378136e3)

Converts state from ENU frame to inertial frame

ThreeBodyProblem.ecef2eci — Function

ecef2eci(rv_ecef, θ; ω=7.292115373194e-5, ang_unit::Symbol=:rad)

Convert state vector rv = [r; v] {km; km/s} from Earth-Centered Earth-Fixed (ECEF) frame to Earth-Centered Inertial (ECI) frame given Greenwich Mean Sidereal Time (GMST) θ {rad}.

ThreeBodyProblem.eci2ecef — Function

ecef2eci(rv_ecef, θ; ω=7.292115373194e-5, ang_unit::Symbol=:rad)

Convert state vector rv = [r; v] {km; km/s} from Earth-Centered Inertial (ECI) frame to Earth-Centered Earth-Fixed (ECEF) frame given Greenwich Mean Sidereal Time (GMST) θ {rad}.

ThreeBodyProblem.eci2sci — Function

eci2sci(rv_eci, rv_sun_eci; ε=23.43929)

ThreeBodyProblem.sci2eci — Function

sci2eci(rv_sci, rv_sun_eci; ε=23.439292)

Missing docstring.

Missing docstring for dimensionalize, dimensionalize!, nondimensionalize, nondimensionalize!. Check Documenter's build log for details.