Orbital Elements

Orbital Elements

Overview

Functions

Overview

The classical Keplerian orbital elements are $[a, e, i Ω, ω, ν]$. While these are the most well-known and commonly used in the astrodynamics community, they are not the only way to describe an orbit. We are also working on including two-line elements and Delaunay variables.

Functions

ThreeBodyProblem.AP2a — Function

AP2a(rA, rP)

Compute the semimajor axis given apoapsis and periapsis distances rA and rP.

ThreeBodyProblem.findrP — Function

computeq(rv::Array,p::Array)

computeq