Parameters

Parameters

Overview

Functions

Overview

The user has control over the parameters defining the primary bodies as well as the system as a whole. We have also included some convenience functions for quickly defining common systems.

Functions

ThreeBodyProblem.Body — Type

Body(m, R, a, T, name, color)
Body(m, R, a, T, name)
Body(m, R, a, T)

A planet, moon, or other gravitationally significant object m::Float64 # mass {kg} R::Float64 # mean radius {km} a::Float64 # mean semimajor axis about parent body {km} T::Float64 # sidereal orbital period {s} name::String # name of body (e.g. "Earth") color::Symbol # color of the body for plotting

ThreeBodyProblem.System — Type

System(prim, sec, μ₁, μ₂, μ, d, R₁, R₂, T, RUNIT, VUNIT, TUNIT, name)
System(prim::Body, sec::Body)

A Circular Restricted Three-Body Problem System defined by primary and secondary bodies (prim and sec) prim::Body # Primary body sec::Body # Secondary body μ₁::Float64 # {km^3/s^2} gravitational parameter of primary body μ₂::Float64 # {km^3/s^2} gravitational parameter of secondary body μ::Float64 # {} mass parameter d::Float64 # {km} average distance between two primaries R₁::Float64 # {km} Radius of primary body R₂::Float64 # {km} Radius of secondary Body T::Float64 # {s} sidereal orbital period RUNIT::Float64 # {km} distance normalizing parameter VUNIT::Float64 # {km/s} velocity normalizing parameter TUNIT::Float64 # {s} time normalizing parameter name::String # name of system (e.g. "Earth/Moon")

Missing docstring.

Missing docstring for set_system. Check Documenter's build log for details.

ThreeBodyProblem.sun_mercury — Function

sun_mercury()

ThreeBodyProblem.sun_venus — Function

sun_venus()

ThreeBodyProblem.sun_earth — Function

sun_earth()

ThreeBodyProblem.earth_moon — Function

earth_moon()

ThreeBodyProblem.sun_mars — Function

sun_mars()

ThreeBodyProblem.sun_jupiter — Function

sun_jupiter()

ThreeBodyProblem.sun_saturn — Function

sun_saturn()

ThreeBodyProblem.sun_uranus — Function

sun_uranus()

ThreeBodyProblem.sun_neptune — Function

sun_neptune()

ThreeBodyProblem.jupiter_europa — Function

jupiter_europa()

ThreeBodyProblem.saturn_enceladus — Function

saturn_enceladus()