Utility functions

Overview

The utility functions are all those functions that didn't fit into the other established categories.

Functions

ThreeBodyProblem.computeR1R2 — Function

 computeR1R2(μ)

Compute the non-dimensional distances of each body from the barycenter given the mass parameter μ {NON}.

 computeR1R2(sys::System)

Compute the non-dimensional distances of each body from the barycenter given CR3BP system sys.

computeR1R2(p::Array)

Compute the dimensional distances of each body from the barycenter given p = [μ₁,μ₂,d] {km³/s², km³/s², km} which are the gravitational parameters of the first and second primary bodies and the distance between them.

ThreeBodyProblem.computer1r2 — Function

computer1r2(rv, μ)

Compute the non-dimensional position vectors of the particle from each body given the state vector rv = [r; v] {NON; NON} and the mass parameter μ {NON}.

computer1r2(rv, sys::System)

Compute the non-dimensional position vectors of the particle from each body given the state vector rv = [r; v] {NON; NON} and the CR3BP system sys.

computer1r2(rv, p::Array)

Compute the dimensional position vectors of the particle from each body given p = [μ₁,μ₂,d] {km³/s², km³/s², km} which are the gravitational parameters of the first and second primary bodies and the distance between them.

ThreeBodyProblem.computeL1 — Function

computeL1(μ; tol=1e-15)

Compute position vector of L1 in a normalized CR3BP given the mass parameter μ {NON}.

computeL1(μ; tol=1e-15)

Compute position vector of L1 in a normalized CR3BP given the CR3BP system sys.

computeL1(p::Array;tol=1e-15)

Compute position vector of L1 in a non-normalized CR3BP given p = [μ₁,μ₂,d] {km³/s², km³/s², km} which are the gravitational parameters of the first and second primary bodies and the distance between them.

ThreeBodyProblem.computeL2 — Function

computeL2(μ; tol=1e-15)

Compute position vector of L2 in a normalized CR3BP given the mass parameter μ {NON}.

computeL2(p::Array;tol=1e-15)

Compute 3D L2 in a non-normalized CR3BP given p = [μ₁,μ₂,d], which are the gravitational parameters of the first and second primary bodies [km³/s²]and the distance between them [km].

ThreeBodyProblem.computeL3 — Function

computeL3(μ;tol=1e-15)

Compute position vector of L3 in a normalized CR3BP given the mass parameter μ {NON}.

computeL3(p::Array;tol=1e-15)

Compute 3D L3 in a non-normalized CR3BP given p = [μ₁,μ₂,d], which are the gravitational parameters of the first and second primary bodies [km³/s²]and the distance between them [km].

ThreeBodyProblem.computeL4 — Function

computeL4(μ; tol=1e-15)

Compute position vector of L4 in a normalized CR3BP given the mass parameter μ {NON}.

computeL4(p::Array;tol=1e-15)

Compute 3D L4 in a non-normalized CR3BP given p = [μ₁,μ₂,d], which are the gravitational parameters of the first and second primary bodies [km³/s²]and the distance between them [km].

ThreeBodyProblem.computeL5 — Function

computeL5(μ;tol=1e-15)

Compute position vector of L5 in a normalized CR3BP given the mass parameter μ {NON}.

computeL5(p::Array;tol=1e-15)

Compute 3D L5 in a non-normalized CR3BP given p = [μ₁,μ₂,d], which are the gravitational parameters of the first and second primary bodies [km³/s²]and the distance between them [km].

ThreeBodyProblem.computeLpts — Function

computeLpts(μ;tol=1e-15)

Compute 3D Lagrange points in a normalized CR3BP given μ, the system mass parameter.

computeLpts(sys::System;tol=1e-15)

Compute 3D Lagrange points in a non-normalized CR3BP given system S.

computeLpts(p::Array;tol=1e-15)

Compute 3D Lagrange points in a non-normalized CR3BP given p = [μ₁,μ₂,d], which are the gravitational parameters of the first and second primary bodies [km³/s²]and the distance between them [km].

ThreeBodyProblem.computeUeff — Function

computeUeff(rv,μ)

Compute effective potential given normalized state rv {NON} and mass parameter μ {NON}.

computeUeff(rv,sys::System)

Compute effective potential given state rv = [r; v]{km; km/s} and CR3BP systemsys`.

computeUeff(rv,p::Array)

Compute effective potential given state rv = [r; v]{km; km/s} andp = [μ₁,μ₂,d]` {km³/s²; km³/s²; km}, which are the gravitational parameters of the first and second primary bodies and the distance between them.

ThreeBodyProblem.computeC — Function

computeC(rv,μ)

Compute Jacobi constant given normalized state rv {NON} and mass parameter {NON}

computeC(rv,sys::System)

Compute Jacobi constant given normalized state rv {NON} and system

computeC(rv,p::Array)

Compute Jacobi constant given state rv = [r; v] {km; km/s} and p = [μ₁,μ₂,d], which are the gravitational parameters of the first and second primary bodies [km³/s²] and the distance between them [km].